Add files via upload
I am sorry.  I am new to GitHub.  Attached is a script which needs to be placed within an Editor folder in your GC2 Unity 6.2+ project.  Via menu, it allows you to add a large number of GC2 Stats and/or Attributes to a GC2 Class asset.  I heavily utilised AI in creating the script.  However, at least it is heavily commented too with explanatory notes.  It should be fairly obvious in it's use.  I've attached some images of it in use as a custom editor tool.